Optimize probing dependencies
