+++ apis/milvus.io/v1beta1/milvus_types.go
@@ -337,7 +337,7 @@ const (
 	ReasonStorageReady       = "StorageReady"
 	ReasonStorageNotReady    = "StorageNotReady"
 	ReasonMsgStreamReady     = "MsgStreamReady"
-	ReasonMsgStreamNotReady  = "MsgStreamReady"
+	ReasonMsgStreamNotReady  = "MsgStreamNotReady"
 	ReasonSecretNotExist     = "SecretNotExist"
 	ReasonSecretErr          = "SecretError"
 	ReasonSecretDecodeErr    = "SecretDecodeError"
