+++ pkg/controllers/milvus_controller_test.go
@@ -6,68 +6,16 @@ import (
 	"time"
 
 	"github.com/golang/mock/gomock"
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	"github.com/milvus-io/milvus-operator/apis/milvus.io/v1beta1"
-	milvusv1beta1 "github.com/milvus-io/milvus-operator/apis/milvus.io/v1beta1"
 	"github.com/milvus-io/milvus-operator/pkg/config"
 	"github.com/milvus-io/milvus-operator/pkg/helm"
 	"github.com/milvus-io/milvus-operator/pkg/util"
 )
-
-var _ = Describe("Milvus controller", func() {
-	const (
-		MCName      = "test"
-		MCNamespace = "default"
-
-		timeout  = time.Second * 10
-		duration = time.Second * 10
-		interval = time.Millisecond * 250
-	)
-
-	Context("Creating Default Milvus", func() {
-		It("It should has default spec", func() {
-			By("By creating a new Milvus")
-			ctx := context.Background()
-			mc := &milvusv1beta1.Milvus{
-				TypeMeta: metav1.TypeMeta{
-					APIVersion: "milvus.io/v1beta1",
-					Kind:       "Milvus",
-				},
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      MCName,
-					Namespace: MCNamespace,
-				},
-			}
-
-			mc.Default()
-			Expect(k8sClient.Create(ctx, mc)).Should(Succeed())
-
-			mcLookupKey := types.NamespacedName{Name: MCName, Namespace: MCNamespace}
-			createdMC := &milvusv1beta1.Milvus{}
-			Eventually(func() bool {
-				err := k8sClient.Get(ctx, mcLookupKey, createdMC)
-				return err == nil
-			}, time.Second*10, interval).Should(BeTrue())
-			// Let's make sure our Schedule string value was properly converted/handled.
-			Expect(*createdMC.Spec.Com.Standalone.Replicas).Should(Equal(int32(1)))
-
-			/* ss := &appsv1.StatefulSetList{}
-			Eventually(func() bool {
-				return nil == k8sClient.List(ctx, ss)
-			}, time.Second*10, interval).Should(BeTrue()) */
-			//Expect(len(ss.Items) > 0).Should(BeTrue())
-
-		})
-	})
-
-})
 
 func TestClusterReconciler(t *testing.T) {
 	config.Init(util.GetGitRepoRootDir())
