+++ pkg/external/kafka.go
@@ -2,9 +2,9 @@ package external
 
 import (
 	"context"
-	"log"
 	"time"
 
+	"github.com/milvus-io/milvus-operator/pkg/util"
 	"github.com/pkg/errors"
 	"github.com/segmentio/kafka-go"
 )
@@ -25,20 +25,5 @@ func CheckKafka(brokerList []string) error {
 		err := r.SetOffsetAt(ctx, time.Now())
 		return errors.Wrap(err, "check consume offset from borker failed")
 	}
-	const backOffInterval = time.Second * 1
-	const maxRetry = 3
-	return doWithBackoff("checkKafka", checkKafka, maxRetry, backOffInterval)
-}
-
-func doWithBackoff(name string, fn func() error, maxRetry int, backOff time.Duration) error {
-	var err error
-	for i := 0; i < maxRetry; i++ {
-		err = fn()
-		if err == nil {
-			return nil
-		}
-		log.Printf("%s with backoff failed, retry %d, err: %v\n", name, i, err)
-		time.Sleep(backOff)
-	}
-	return errors.Wrap(err, name+" with backoff failed")
+	return util.DoWithBackoff("checkKafka", checkKafka, util.DefaultMaxRetry, util.DefualtBackOffInterval)
 }
